Update default resources limits and requests in values.yaml
In this update, default values for cpu and memory limits and requests have been uncommented and specified under resources field for both primary and worker in charts/chatwoot/values.yaml, to ensure efficient use of resources. Also, service.beta.kubernetes.io/aws-load-balancer-type annotation was moved in line with others and readOnlyRootFilesystem was made true for added security. Finally, the version was bumped to 1.0.77 in charts/chatwoot/Chart.yaml.