Update version in Chart.yaml
The Chart.yaml file in the Chatwoot folder was updated to reflect the new version of the application (1.0.79). This change is part of a larger update that includes amendments to default resource limits in values.yaml, adjustments to annotations, and enhanced security measures. All of these amendments were necessary to effectively manage resources, maintain security, and keep all our