Update minpods, replicaCount and memory limits in values.yaml
The changes in charts/chatwoot/values.yaml have reduced the minimum number of pods for both server and worker from 4 and 8 to 1 and 2 respectively. The count of replica has also been updated from 4 and 8 to 1 and 2. These modifications are made to better manage system resources.

The memory limit for both server and worker has been increased from 512Mi to 2Gi and requested memory from 256Mi to 1Gi. This change has been made since the previous setting was too low for the service to function properly with increased load.

Chart version is updated to 1.0.80 following these modifications.