Update Chatwoot image and chart version
The image tag in 'values.yaml' and the chart version in 'Chart.yaml' have been updated. The changes were prompted by a new release of the Chatwoot application, necessitating the update of the image tag from v1.0.30-ce-telelogx to v1.0.31-ce-telelogx and the increment of the chart version to v1.0.83.